requirements.txt
requests==2.31.0
pymysql
cryptography
python-dotenv==1.0.0